Screen: always update & monitor for resize events
Removing the monitorResize function, since anyone using this plugin would likely be doing so to always keep the navbar hidden.

Also adding the update function to the start of toggleSystemUI function so that the callback function will get called for all devices and platforms when the app launches or resumes.

Also adding a screen.callbackCalled property that can be checked for if the callback function has been called. This could be used for cases where there is significant and/or unforeseeable delay between requiring the screen module and actually needing to use the callback function.