Performance: improve performance & add average FPS option
Greatly improve performance by utilising table.concat to create the performance meter counter text element and remove the forced garbage collection.

Furthermore, adding an an optional "framesBetweenUpdate" (default value is 0) that can be used to determine how frequently the performance meter is updated.

Also adjusting the default visual parameters to center the performance meter to the top of the screen by default and making it slightly transparent.