Print To Display: add a safety check when exceeding max texture size
Especially on mobile devices, if the user tries to print massive strings, such sending a network request to a page and then trying to print out the entire event.response, then the device may run out of texture memory.

If the perceived max texture size limit is exceeded, then print out a warning and a stub of the intended print instead.