Print-To-Display: add support for any length outputs
The previous version of Print to Display only prompted the user that an output was too long, but the new version will display a console ouput of any length.

Also added a sample custom button to the Print To Display sample project.

Also, writing this on a new Mac and its Atom installation wiped all extra whitespace from the end of lines.