Update UUID.lua
Making the UUID entirely time based to avoid issues with deviceID (with newer Android versions). This makes the module work better with HTML5 platform as well, though sockets still aren't available for it.

Finally, there's now an option to hash the UUID, making it perfectly anonymous and uniform in length.