Print to Display: refactor unhandledError logging
The new errorHandling parameters can be passed in a table, just like the custom style parameters would be.

These errorHandling parameters control how the console will react to encountering runtime errors, i.e. whether it'll react to them or not, and if it does, then will it print out the error before the crash, will it open to show the error if it is closed at the time, and will it try to suppress the error or not.

The console's drag controls are now also smoother for toggling autoscroll on or off, leaving 10px area at the bottom that will trigger the autoscroll on if the scroll is dragged there.