Print to Display: add new functionalities
Added option to launch the console with the controls hidden (controls are visible by default).

Added the ability to scroll the console using the mouse wheel when hovering over the console on desktop (on by default).

Improved the scroll capping to boundaries (wasn't particularly visible with just touch controls).

Added a convenience function to clear the console programmatically.